refactor: hook을 하나로 파일로 관리하여 중복 제거
