+++ src/components/store/ShopCardsShelf.js
@@ -1,8 +1,7 @@
-import React, { useState, useEffect } from "react";
 import { Swiper, SwiperSlide } from "swiper/react";
 import "swiper/swiper-bundle.css";
+import { useResizeHandler } from "../hooks/useResizeHandler";
 
-// Product card data (based on the HTML structure provided)
 const products = [
   {
     title: "iPhone 16 Pro",
@@ -106,36 +105,7 @@ const products = [
   },
 ];
 const ShopCardsShelf = () => {
-  const [containerStyle, setContainerStyle] = useState({});
-  const [slidesOffsetBefore, setSlidesOffsetBefore] = useState(0);
-
-  useEffect(() => {
-    const handleResize = () => {
-      const screenWidth = window.innerWidth;
-      let calculatedMargin;
-      let calculatedOffset;
-
-      if (screenWidth > 1440) {
-        calculatedMargin = "140px";
-        calculatedOffset = 140;
-      } else if (screenWidth >= 1024 && screenWidth <= 1440) {
-        calculatedMargin = `calc(-268.46154px + 28.36538vw)`;
-        const offsetValue = screenWidth * 0.28 - 268.46154;
-        calculatedOffset = Math.max(20, offsetValue);
-      } else {
-        calculatedMargin = "20px";
-        calculatedOffset = 20;
-      }
-
-      setContainerStyle({ marginLeft: calculatedMargin });
-      setSlidesOffsetBefore(calculatedOffset);
-    };
-
-    window.addEventListener("resize", handleResize);
-    handleResize();
-
-    return () => window.removeEventListener("resize", handleResize);
-  }, []);
+  const { containerStyle, slidesOffsetBefore, isReady } = useResizeHandler();
 
   return (
     <>
